requirements.txt
streamlit==1.35.0
yt-dlp==2024.11.18
openai-whisper==20240930
spacy==3.7.5
pandas==2.0.2
numpy==1.26.4
pyarrow==17.0.0
